+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.5.0] - 2023-12-10
+### Added
+- Added a `search` command
+### Changed
+- Changed the style of the progressbar for status
+
 ## [0.4.1] - 2023-09-17
 ### Added
 - Added a `--destination` flag
@@ -51,7 +57,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Refactoring of all code
 
-[Unreleased]: https://github.com/flowrey/youtube-bz/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/flowrey/youtube-bz/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/flowrey/youtube-bz/releases/tag/v0.5.0
 [0.4.1]: https://github.com/flowrey/youtube-bz/releases/tag/v0.4.1
 [0.4.0]: https://github.com/flowrey/youtube-bz/releases/tag/v0.4.0
 [0.3.4]: https://github.com/flowrey/youtube-bz/releases/tag/v0.3.4